requirements.txt
python-dotenv
PyTelegramBotAPI
requests